pyproject.toml
[tool.poetry]
name = "music-streamer"
version = "0.1.0"
description = ""
authors = ["tavkhelidzeluka <tavkhelidzeluka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
pillow = "^10.3.0"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"
audioread = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
